# Trabalho_API/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Trabalho_API",
        "version": "1.0.0",
        "description": "API para cadastro e edição de carros de F1"
    },
    "paths": {
        "/carros": {
            "get": {
                "summary": "Obter a lista de carros",
                "responses": {
                    "200": {
                        "description": "Lista de retornada com sucesso"
                    }
                }
            },
            "post": {
                "summary": "Criar um novo carro",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carros"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Carro criado com sucesso"
                    },
                    "400": {
                        "description": "Dados inválidos fornecidos"
                    }
                }
            }
        },
        "/carros/{id}": {
            "get": {
                "summary": "Obter informações de um carro por ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do carro retornadas com sucesso"
                    },
                    "404": {
                        "description": "Carro não encontrado"
                    }
                }
            },
            "put": {
                "summary": "Atualizar informações de um carro pelo ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Carros"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Informações do carro atualizadas com sucesso"
                    },
                    "404": {
                        "description": "Carro não encontrado"
                    }
                }
            },
            "delete": {
                "summary": "Excluir um carro pelo ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Carro excluído com sucesso"
                    },
                    "404": {
                        "description": "Carro não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Carros": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "pilots": {
                        "type": "string"
                    },
                    "team": {
                        "type": "string"
                    },
                    "img": {
                        "type": "string"
                    }
                }
            }
        }
    }
}